bulk rename getball-fetch to getball_fetch, according to new no-dash component naming design rule, see commit b553554
