+++ components/falcons/getball/dependencies
@@ -1,2 +1,2 @@
-components/falcons/getball-fetch
+components/falcons/getball_fetch
 components/falcons/getball-intercept
