+++ components/falcons/getball/interface/BUILD
@@ -11,7 +11,7 @@ proto_library(
     deps = [
         "@com_google_protobuf//:timestamp_proto",
         "//datatypes:MRA_proto",
-        "//components/falcons/getball-fetch/interface:interface_proto",
+        "//components/falcons/getball_fetch/interface:interface_proto",
         "//components/falcons/getball-intercept/interface:interface_proto",
     ],
 )
