+++ components/falcons/getball_fetch/BUILD
@@ -8,14 +8,14 @@ cc_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//components/robotsports/getball-fetch/interface:interface",
+        "//components/falcons/getball_fetch/interface:interface",
     ],
 )
 
 cc_library(
     name = "implementation",
     hdrs = [
-        "RobotsportsGetballFetch.hpp", # generated by MRA-codegen.py
+        "FalconsGetballFetch.hpp", # generated by MRA-codegen.py
     ],
     srcs = [
         "tick.cpp", # to be implemented by user, generated if not existing
