+++ components/falcons/getball_fetch/datatypes.hpp
@@ -4,9 +4,9 @@
 #ifndef _MRA_FALCONS_GETBALL_FETCH_DATATYPES_HPP
 #define _MRA_FALCONS_GETBALL_FETCH_DATATYPES_HPP
 
-#include "components/falcons/getball-fetch/interface/Input.pb.h"
-#include "components/falcons/getball-fetch/interface/Params.pb.h"
-#include "components/falcons/getball-fetch/interface/Output.pb.h"
+#include "components/falcons/getball_fetch/interface/Input.pb.h"
+#include "components/falcons/getball_fetch/interface/Params.pb.h"
+#include "components/falcons/getball_fetch/interface/Output.pb.h"
 
 
 #endif
