+++ components/robotsports/getball_fetch/testdata/kickoff_prepare.json
